update
- update folder structure
- load factory directory and suffic from rails config
- properly load rails env
