+++ lib/model_2_factory-rails.rb
@@ -1,0 +1,2 @@
+require 'model_2_factory/model_2_factory'
+require 'model_2_factory/version'
