--- lib/model_2_factory/rails.rb
@@ -1,1 +1,0 @@
-require 'rails/version'
