--- lib/model_2_factory/rails/version.rb
@@ -1,7 +1,0 @@
-module Model2Factory
-  module Rails
-    module Version
-      STRING = '0.0.1'.freeze
-    end
-  end
-end
