+++ lib/model_2_factory/version.rb
@@ -1,0 +1,3 @@
+module Model2Factory
+  VERSION = '0.0.1'.freeze
+end
