--- lib/model_2_factory_rails.rb
@@ -1,1 +1,0 @@
-require 'model_2_factory/rails/model_2_factory'
